move CLI logic under RipperTags and invoke it explicitly
Before, the bare act of `require "ripper-tags"` would process ARGV and
start generating tags. This is unwanted and potentially destructive
behavior if the library is loaded in an application, e.g. added to
Gemfile of a Rails app.

This moves all CLI logic to methods under RipperTags and requires the
`ripper-tags` executable to trigger the behavior explicitly. Bonus
side-effect is that the whole thing is now testable.