fix superclass detection on Ruby 1.8
In case of `Class.new klass`, we don't want `klass` being interpreted as
a name of superclass since it's not a constant.

The unit test for this failed on Ruby 1.8 since the backport of Ripper
parses that line differently than its 1.9 equivalent; it interprets
`klass` as `var_ref` rather than `vcall`. This was tricky to debug due
to lsegal/ripper18#7

This quick workaround simply disallows superclass names that don't look
like a Ruby constant.