On parse errors, only exit with error status if all files have failed
A parse error in one or few files out of many should not halt the
process, since ripper-tags typically handles many files at once.
However, if only one file was given, and that file fails, have the
process exit with status 1 to indicate that the tags output is likely
useless.